chore: add menu-close and menu-open SVG icons
This commit adds two new SVG icons, menu-close.svg and menu-open.svg, to the public/icons directory.
The menu-close.svg icon is used to represent a closed menu, while the menu-open.svg icon is used to
represent an open menu.

These icons will be used in the application's navigation menu to provide visual feedback to the user
when the menu is open or closed.